refactor: merge duplicate code to populateDraftAccounts func
