+++ src/renderer/features/wallets/KeyConstructor/index.ts
@@ -1,2 +1,3 @@
 export { KeyConstructor } from './ui/KeyConstructor';
 export { type DerivationKeyDraft } from './model/constructor-model';
+export { populateDraftAccounts } from './lib/utils';
